feat: add support for reporterAPI usage
This makes the reporter config the default and deprecates testResultsProcessor.

Fixes #32

BREAKING CHANGE: Reporter API is the default, testResultsProcessor is deprecated.
