+++ examples/scripts/jestTrxProcessor.js
@@ -1,7 +1,7 @@
-var builder = require('jest-trx-results-processor');
+var builder = require("jest-trx-results-processor/dist/testResultsProcessor");
 
 var processor = builder({
-    outputFile: 'relative/path/to/resulting.trx' // this defaults to "test-results.trx"
+  outputFile: "relative/path/to/resulting.trx", // this defaults to "test-results.trx"
 });
 
 module.exports = processor;
